tetragon: Move kprobe constants to consts.go
Add constants specified in generickprobe.go into constants_linux.go
This allows generickprobe.go to compile on Windows
Plus removing enforcerDataMapName, maxStringSizeSmall, maxStringSizeTiny
