+++ pkg/sensors/tracing/generickprobe.go
@@ -58,33 +58,6 @@ func init() {
 	observer.RegisterEventHandlerAtInit(ops.MSG_OP_GENERIC_KPROBE, handleGenericKprobe)
 }
 
-const (
-	CharBufErrorENOMEM      = -1
-	CharBufErrorPageFault   = -2
-	CharBufErrorTooLarge    = -3
-	CharBufSavedForRetprobe = -4
-
-	// The following values could be fine tuned if either those feature use too
-	// much kernel memory when enabled.
-	stackTraceMapMaxEntries = 32768
-	ratelimitMapMaxEntries  = 32768
-	fdInstallMapMaxEntries  = 32000
-	enforcerMapMaxEntries   = 32768
-	overrideMapMaxEntries   = 32768
-)
-
-func kprobeCharBufErrorToString(e int32) string {
-	switch e {
-	case CharBufErrorENOMEM:
-		return "CharBufErrorENOMEM"
-	case CharBufErrorTooLarge:
-		return "CharBufErrorBufTooLarge"
-	case CharBufErrorPageFault:
-		return "CharBufErrorPageFault"
-	}
-	return "CharBufErrorUnknown"
-}
-
 type kprobeSelectors struct {
 	entry *selectors.KernelSelectorState
 	retrn *selectors.KernelSelectorState
@@ -1193,11 +1166,6 @@ func loadGenericKprobeSensor(bpfDir string, load *program.Program, maps []*progr
 	return fmt.Errorf("invalid loadData type: expecting idtable.EntryID/[] and got: %T (%v)",
 		load.LoaderData, load.LoaderData)
 }
-
-var errParseStringSize = errors.New("error parsing string size from binary")
-
-// this is from bpf/process/types/basic.h 'MAX_STRING'
-const maxStringSize = 4096
 
 func getUrl(url string) {
 	// We fire and forget URLs, and we don't care if they hit or not.
